requirements.txt
git+https://github.com/OCA/openupgradelib.git
git-aggregator
copier
pre-commit
debugpy
pydevd-odoo
pylint_odoo
invoke
ensurepath
PyYAML
ruff
black
types-invoke
openpyxl-stubs
pandas
